Combine navigation-list, navigation-edit, and navigation-shared into one navigation route
