--- routes/navigation-shared/package.json
@@ -1,11 +1,0 @@
-{
-	"name": "@wordpress/navigation-shared",
-	"version": "1.0.0",
-	"private": true,
-	"dependencies": {
-		"@wordpress/blocks": "file:../../packages/blocks",
-		"@wordpress/core-data": "file:../../packages/core-data",
-		"@wordpress/element": "file:../../packages/element",
-		"@wordpress/lazy-editor": "file:../../packages/lazy-editor"
-	}
-}
