+++ routes/navigation/package.json
@@ -7,6 +7,23 @@
 		"page": "site-editor-v2"
 	},
 	"dependencies": {
-		"@wordpress/route": "file:../../packages/route"
+		"@wordpress/admin-ui": "file:../../packages/admin-ui",
+		"@wordpress/base-styles": "file:../../packages/base-styles",
+		"@wordpress/block-editor": "file:../../packages/block-editor",
+		"@wordpress/block-library": "file:../../packages/block-library",
+		"@wordpress/blocks": "file:../../packages/blocks",
+		"@wordpress/components": "file:../../packages/components",
+		"@wordpress/core-data": "file:../../packages/core-data",
+		"@wordpress/data": "file:../../packages/data",
+		"@wordpress/dataviews": "file:../../packages/dataviews",
+		"@wordpress/editor": "file:../../packages/editor",
+		"@wordpress/element": "file:../../packages/element",
+		"@wordpress/html-entities": "file:../../packages/html-entities",
+		"@wordpress/i18n": "file:../../packages/i18n",
+		"@wordpress/icons": "file:../../packages/icons",
+		"@wordpress/lazy-editor": "file:../../packages/lazy-editor",
+		"@wordpress/notices": "file:../../packages/notices",
+		"@wordpress/route": "file:../../packages/route",
+		"@wordpress/views": "file:../../packages/views"
 	}
 }
